feat: 添加房源系统事件处理

+++ domus/src/domain/house/events/house_updated.rs
@@ -46,14 +46,14 @@ pub struct HouseUpdatedEvent {
     // 级别
     pub level: Option<String>,
     // 层高
-    pub floor_height: Option<f64>,
+    pub floor_height: Option<f32>,
     // 进深
     pub progress_depth: Option<f64>,
     // 门宽
     pub door_width: Option<f64>,
 
     // 使用面积
-    pub use_area: Option<f64>,
+    pub use_area: Option<f32>,
     // 售价
     pub sale_price: Option<f64>,
     // 租价
@@ -78,7 +78,7 @@ pub struct HouseUpdatedEvent {
     // 建筑结构
     pub building_structure: Option<String>,
     // 建筑年代
-    pub building_year: Option<u32>,
+    pub building_year: Option<i32>,
     // 产权性质
     pub property_rights: Option<String>,
     // 产权年限
