+++ domus/src/domain/house/value_objects/house_update_data.rs
@@ -48,14 +48,14 @@ pub struct HouseUpdateData {
     // 级别
     pub level: Option<String>,
     // 层高
-    pub floor_height: Option<f64>,
+    pub floor_height: Option<f32>,
     // 进深
     pub progress_depth: Option<f64>,
     // 门宽
     pub door_width: Option<f64>,
 
     // 使用面积
-    pub use_area: Option<f64>,
+    pub use_area: Option<f32>,
     // 售价
     pub sale_price: Option<f64>,
     // 租价
@@ -80,7 +80,7 @@ pub struct HouseUpdateData {
     // 建筑结构
     pub building_structure: Option<String>,
     // 建筑年代
-    pub building_year: Option<u32>,
+    pub building_year: Option<i32>,
     // 产权性质
     pub property_rights: Option<String>,
     // 产权年限
